fix: Make cloud crypto function, update flow

+++ cloud/api.py
@@ -4,12 +4,12 @@ import logging
 import pathlib
 from datetime import datetime
 from json import JSONDecodeError
-from typing import Literal, Annotated
+from typing import Literal, Annotated, Optional
 
 from Crypto.PublicKey import ECC
 from functools import lru_cache
 from fastapi import FastAPI, Depends, Path, Body
-from fastapi.openapi.models import Example
+from fastapi.params import Query
 from fastapi.responses import JSONResponse
 from tinydb import TinyDB, where
 from tinydb.table import Document
@@ -17,9 +17,10 @@ from tinydb.table import Document
 from crypto.crypto import load_private_key, generate_keys, encrypt_hpke, determine_role, \
     decrypt_hpke, verify_vc
 from dotenv import load_dotenv
-from util.models import EncryptedPayload, SuccessfulResponse, EncryptedPayloadVC, DID, \
-    BadRequestResponse, ForbiddenResponse, NotFoundResponse, VerifiableCredential
+from util.models import EncryptedPayload, SuccessfulResponse, DID, \
+    BadRequestResponse, ForbiddenResponse, NotFoundResponse, VerifiablePresentation, get_encrypted_payload
 from util.middleware import verify_request
+from util.validators import validate_battery_pass_payload
 
 app = FastAPI(
     title="Battery Pass API",
@@ -43,10 +44,10 @@ with open(pathlib.Path(__file__).parent.parent /
     vc_service_json = json.load(f)
 
 
-def example_payload(did: DID, vc: VerifiableCredential):
+def example_payload(did: DID, vp: VerifiablePresentation):
     payload = payload_json.copy()
     payload["did"] = did
-    payload["vc"] = vc
+    payload["vp"] = vp
     return payload
 
 
@@ -78,7 +79,10 @@ def set_nested_value(doc, path_keys, new_value):
     current_level = doc
     for key in path_keys[:-1]:
         current_level = current_level.setdefault(key, {})
-    current_level[path_keys[-1]] = new_value
+    if isinstance(current_level[path_keys[-1]], list):
+        current_level[path_keys[-1]].append(new_value)
+    else:
+        current_level[path_keys[-1]] = new_value
 
 
 @app.get("/",
@@ -111,6 +115,9 @@ def retrieve_data(scope: Literal["public", "bms", "legitimate_interest"], did: s
         private_key=private_key,
         bundle=doc["encrypted_data"]
     )
+    if scope == "bms":
+        return json.loads(decrypted_data)
+    return None
 
 
 @app.get("/batterypass/{did}",
@@ -122,43 +129,37 @@ def retrieve_data(scope: Literal["public", "bms", "legitimate_interest"], did: s
              404: {"model": NotFoundResponse},
          })
 async def read_item(
-        did: DID,
-        item: Annotated[EncryptedPayloadVC, Body(openapi_examples={
-            "public": {
-                "summary": "Public scope",
-                "description": "Requires an empty body.",
-                "value": {}
-            },
-            "bms": {
-                "summary": "BMS scope",
-                "description": "Requires a signature signed by the BMS.<br><br>"
-                               "The payload inside the ciphertext must be a **128-byte random number**.",
-                "value": example_payload(did="did:batterypass:bms.sn-987654321", vc=None)
-            },
-            "vc": {
-                "summary": "VC-defined scope",
-                "description": "Requires a verifiable credential signed by the BMS defining the access level.<br><br>"
-                               "The payload inside the ciphertext must be a **128-byte random number**.",
-                "value": example_payload(did="did:batterypass:bms.sn-987654321", vc=vc_service_json)
-            }
-        })],
+        did: str,
+        payload: str = Query(
+            default=None,
+            description="An [encrypted JSON payload](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                        "/blob/main/cloud/docs/api.md#request-body) as a serialized string.\n\n"
+                        "The payload inside the ciphertext must contain a 128-byte random number."
+        ),
+        public: bool = Query(default=True, description="Whether to retrieve only public battery pass data.\n\n"
+                                                       "If set to `false`, either `payload` or `vp` must be provided."),
+        vp: str = Query(default=None,
+                        description="A verifiable presentation signed by the BMS defining the access level as a "
+                                    "serialized JSON string."),
         db: TinyDB = Depends(get_db),
         private_key: ECC.EccKey = Depends(get_private_key),
 ):
     document = db.search(where("did") == did)
     if not document:
         return error_response(404, "Entry doesn't exist.")
-    if not item:
+    if public:
         return retrieve_data(scope="public", did=did, doc=document[0], private_key=private_key)
+    vp: VerifiablePresentation = VerifiablePresentation.model_validate_json(vp) if vp else None
+    payload: EncryptedPayload = EncryptedPayload.model_validate_json(payload) if payload else None
     try:
-        random_number = verify_request(item, private_key)
+        random_number = verify_request(payload, private_key)
         if len(random_number) != 128:
             raise ValueError("Invalid length for random value.")
     except ValueError as e:
         return error_response(400, e.args[0])
-    if determine_role(db, did, item["did"]) == "bms":
+    if determine_role(document[0], payload.did) == "bms":
         return retrieve_data(scope="bms", did=did, doc=document[0], private_key=private_key)
-    if "vc" in item and verify_vc(item["vc"]):
+    if vp and verify_vc(vp):
         return retrieve_data(scope="legitimate_interest", did=did, doc=document[0], private_key=private_key)
     return error_response(400, "Invalid request.")
 
@@ -175,33 +176,32 @@ async def read_item(
                    }},
              400: {"model": BadRequestResponse},
              403: {"model": ForbiddenResponse},
-         })
+         },
+         description="A detailed description can be found "
+                     "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                     "/blob/main/cloud/docs/api.md#put-batterypassdid)**.")
 async def create_item(
-        item: Annotated[EncryptedPayload, Body(openapi_examples={
-            "default": {
-                "summary": "Default",
-                "description": "A detailed description can be found "
-                               "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data/blob/main/cloud/docs/api.md#put-batterypassdid)**.",
-                "value": example_payload(did="did:batterypass:bms.sn-987654321", vc=None)
-            }
-        })],
-        did: str = Path(description="A properly formed DID"),
+        payload: EncryptedPayload,
+        did: str,
         db: TinyDB = Depends(get_db),
         private_key: ECC.EccKey = Depends(get_private_key),
 ):
+    logging.info("Test")
     try:
-        decrypted_item = verify_request(item, private_key)
+        decrypted_payload = verify_request(payload, private_key)
     except ValueError as e:
         return error_response(400, e.args[0])
-
-    if not determine_role(db, did, item["did"]) == "oem":
-        return error_response(403, "Access denied.")
     if db.search(where("did") == did):
         logging.warning(f"DID {did} already exists in DB")
         return error_response(400, "Entry already exists.")
+    if not determine_role(None, payload.did) == "oem":
+        return error_response(403, "Access denied.")
+    results = validate_battery_pass_payload(json.loads(decrypted_payload))
+    if not all(value == "Valid" for value in results.values()):
+        return error_response(400, f"Invalid payload: {json.dumps(results)}")
     db.insert({
         "did": did,
-        "encrypted_data": encrypt_hpke(private_key, decrypted_item)
+        "encrypted_data": encrypt_hpke(private_key, decrypted_payload)
     })
     return {"ok": f"Entry for {did} added successfully."}
 
@@ -219,34 +219,30 @@ async def create_item(
               400: {"model": BadRequestResponse},
               403: {"model": ForbiddenResponse},
               404: {"model": NotFoundResponse},
-          })
+          },
+          description="A detailed description can be found "
+                      "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                      "/blob/main/cloud/docs/api.md#post-batterypassdid)**.")
 async def update_item(
-        item: Annotated[EncryptedPayload, Body(openapi_examples={
-            "default": {
-                "summary": "Default",
-                "description": "A detailed description can be found "
-                               "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data/blob/main/cloud/docs/api.md#post-batterypassdid)**.",
-                "value": example_payload(did="did:batterypass:bms.sn-987654321", vc=None)
-            }
-        })],
-        did: str = Path(description="A properly formed DID"),
+        payload: EncryptedPayload,
+        did: str,
         db: TinyDB = Depends(get_db),
         private_key: ECC.EccKey = Depends(get_private_key),
 ):
     try:
-        decrypted_item = json.loads(verify_request(item, private_key))
+        decrypted_payload = json.loads(verify_request(payload, private_key))
     except JSONDecodeError:
         return error_response(400, "Error occurred while decoding JSON.")
     except ValueError as e:
         return error_response(400, e.args[0])
 
     document = db.search(where("did") == did)
-    if determine_role(db, did, item["did"]) != "bms":
-        return error_response(403, "Access denied.")
     if not document:
         return error_response(404, "Entry doesn't exist.")
+    if determine_role(document[0], payload.did) != "bms":
+        return error_response(403, "Access denied.")
     decrypted_document = json.loads(decrypt_hpke(private_key, document[0]["encrypted_data"]))
-    for element in decrypted_item:  # Iterate over the list of JSON items
+    for element in decrypted_payload:  # Iterate over the list of JSON items
         if not isinstance(element, dict) or len(element) != 1:
             return error_response(400, "Invalid update format.")
         key, value = next(iter(element.items()))
@@ -271,27 +267,25 @@ async def update_item(
                 400: {"model": BadRequestResponse},
                 403: {"model": ForbiddenResponse},
                 404: {"model": NotFoundResponse},
-            })
+            },
+            description="A detailed description can be found "
+                        "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                        "/blob/main/cloud/docs/api.md#delete-batterypassdid)**.")
 async def delete_item(
-        item: Annotated[EncryptedPayload, Body(openapi_examples={
-            "default": {
-                "summary": "Default",
-                "description": "A detailed description can be found "
-                               "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data/blob/main/cloud/docs/api.md#delete-batterypassdid)**.",
-                "value": example_payload(did="did:batterypass:bms.sn-987654321", vc=None)
-            }
-        })],
-        did: str = Path(description="A properly formed DID"),
+        did: str,
+        payload: str = Query(
+            description="An [encrypted JSON payload](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                        "/blob/main/cloud/docs/api.md#request-body) as a serialized string.\n\n"
+                        "The payload inside the ciphertext must contain a 128-byte random number."
+        ),
         db: TinyDB = Depends(get_db),
         private_key: ECC.EccKey = Depends(get_private_key),
 ):
+    payload = EncryptedPayload.model_validate_json(payload)
     try:
-        verify_request(item, private_key)
+        verify_request(payload, private_key)
     except ValueError as e:
         return error_response(400, e.args[0])
-
-    if determine_role(db, did, item["did"]) != "bms":
-        return error_response(403, "Access denied.")
 
     # Search for database entry with the given DID
     document = db.search(where("did") == did)
@@ -299,6 +293,10 @@ async def delete_item(
     # If no entry is found, raise an HTTP exception
     if not document:
         return error_response(404, "Entry doesn't exist.")
+
+    if determine_role(document[0], payload.did) != "bms":
+        return error_response(403, "Access denied.")
+
 
     # Delete the entry from the database
     db.remove(where("did") == did)
